fix: rename `cover_crypt` file into `primitives`
